requirements.txt
torch
torchvision
pytorch-lightning
gradio
opencv-python
PyYAML
Pillow
numpy
scikit-learn
tensorboard
gunicorn
torchcam
matplotlib
